requirements.txt
asgiref==3.5.2
Django==4.0.1
sqlparse==0.4.3
tzdata==2022.6
django-environ==0.9.0
psycopg2-binary==2.9.5
djangorestframework==3.14.0
django-cors-headers==3.13.0
social-auth-app-django==5.0.0
django-filter==22.1
marshmallow==3.14.0
marshmallow-dataclass==8.5.10
requests==2.28.1
pytest==7.2.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1